requirements.txt
streamlit==1.32.2
openai==0.27.8
langchain==0.0.325
faiss-cpu
pdfplumber
python-dotenv
deep-translator
langdetect